more more more testing but i'm passing 51 tests

+++ api/quotes/delete.php
@@ -12,6 +12,7 @@
 
   include_once '../../config/Database.php';
   include_once '../../models/Quote.php';
+  include_once '../../functions/isValid.php';
 
   // Instantiate DB & connect
   $database = new Database();
@@ -27,7 +28,13 @@
 
   $quote->id = $data->id;
   
-
+  $idExists = isValid($id, $quote);
+  if (!$idExists) { 
+      echo json_encode(
+          array('message' => 'No Quotes Found')
+      );
+      exit();
+  }
   //Delete quote
   if($quote->delete()) {
     echo json_encode($quote->delete());
